v1/requirements.txt
# requirements.txt
# python 3.11.11
langgraph==0.3.16
langchain-core==0.3.55
langchain-openai==0.3.9
langchain-community==0.3.20
langchain-text-splitters==0.3.7
langchain-teddynote==0.3.44
selenium==4.31.0
beautifulsoup4==4.13.3
requests==2.32.3
python-dotenv==1.0.1
faiss-cpu==1.10.0
google-cloud-aiplatform==1.90.0
pydantic==2.10.6
typing-extensions==4.12.2
selenium-stealth==1.0.6
webdriver-manager==4.0.2
easyocr==1.7.2
fastapi==0.115.12
uvicorn==0.34.2
pytest==8.3.5
pytest-mock